+++ src/components/card/card.stories.tsx
@@ -15,7 +15,9 @@ export const Default: Story = {
     children: (
       <div className="flex flex-col gap-2">
         <h4 className="font-bold text-neutral-1 text-base">Card title</h4>
-        <p className="text-xs text-neutral-2">Explanatory card content within the interface.</p>
+        <p className="text-xs text-muted-on-light">
+          Explanatory card content within the interface.
+        </p>
       </div>
     ),
   },
@@ -26,7 +28,9 @@ export const Playground: Story = {
     children: (
       <div className="flex flex-col gap-2">
         <h4 className="font-bold text-neutral-1 text-base">Card title</h4>
-        <p className="text-xs text-neutral-2">Explanatory card content within the interface.</p>
+        <p className="text-xs text-muted-on-light">
+          Explanatory card content within the interface.
+        </p>
       </div>
     ),
   },